cleaned the whole parameter parsing system, removed redundancies
