+++ src/main/kotlin/com/papsign/ktor/openapigen/modules/openapi/HandlerModule.kt
@@ -3,7 +3,6 @@ package com.papsign.ktor.openapigen.modules.openapi
 import com.papsign.ktor.openapigen.OpenAPIGen
 import com.papsign.ktor.openapigen.modules.ModuleProvider
 import com.papsign.ktor.openapigen.modules.OpenAPIModule
-import com.papsign.ktor.openapigen.parameters.parsers.ParameterHandler
 
 interface HandlerModule: OpenAPIModule {
     fun configure(apiGen: OpenAPIGen, provider: ModuleProvider<*>)
