+++ src/main/kotlin/com/papsign/ktor/openapigen/parameters/QueryParamStyle.kt
@@ -1,5 +1,5 @@
 package com.papsign.ktor.openapigen.parameters
 
 enum class QueryParamStyle: ParameterStyle {
-    form, spaceDelimited, pipeDelimited, deepObject
+    DEFAULT, form, spaceDelimited, pipeDelimited, deepObject
 }
