+++ src/main/kotlin/com/papsign/ktor/openapigen/parameters/handlers/ParameterHandler.kt
@@ -1,0 +1,7 @@
+package com.papsign.ktor.openapigen.parameters.handlers
+
+import io.ktor.http.Parameters
+
+interface ParameterHandler<T> {
+    fun parse(parameters: Parameters): T
+}
