--- src/main/kotlin/com/papsign/ktor/openapigen/parameters/parsers/NoTranslation.kt
@@ -1,6 +1,0 @@
-package com.papsign.ktor.openapigen.parameters.parsers
-
-object NoTranslation : OpenAPIPathSegmentTranslator {
-    override fun translateSegment(segment: String): String = segment
-    override fun getQuerySegment(): String = ""
-}
