--- src/main/kotlin/com/papsign/ktor/openapigen/parameters/parsers/PathParameterTranslation.kt
@@ -1,7 +1,0 @@
-package com.papsign.ktor.openapigen.parameters.parsers
-
-import com.papsign.ktor.openapigen.parameters.PathParamStyle
-
-data class PathParameterTranslation(val key: String, val style: PathParamStyle, val explode: Boolean) {
-    val translatedName = "${style.prefix}$key${if (explode) "*" else ""}"
-}
