+++ src/main/kotlin/com/papsign/ktor/openapigen/parameters/parsers/UnitParameterHandler.kt
@@ -1,0 +1,9 @@
+package com.papsign.ktor.openapigen.parameters.parsers
+
+import com.papsign.ktor.openapigen.parameters.handlers.ParameterHandler
+import io.ktor.http.Parameters
+
+object UnitParameterHandler :
+    ParameterHandler<Unit> {
+    override fun parse(parameters: Parameters) = Unit
+}
